size_t is not portable, using it broke the selection feature. replace every size_t to unsigned int.

+++ Camera/BackCamera.cpp
@@ -66,15 +66,15 @@ void BackCamera::setCameraForSelectionS()
     glViewport((GLint)m_startX,(GLint)m_startY,(GLint)m_width,(GLint)m_height);
 }
 
-void BackCamera::setCameraForSelectionD(size_t x1,size_t y1,size_t x2,size_t y2,size_t h)
+void BackCamera::setCameraForSelectionD(unsigned int x1,unsigned int y1,unsigned int x2,unsigned int y2,unsigned int h)
 {
     glViewport((GLint)m_startX,(GLint)m_startY,(GLint)m_width,(GLint)m_height);
     GLint viewport[4];
     glGetIntegerv (GL_VIEWPORT, viewport);
     glMatrixMode(GL_PROJECTION);
     glLoadIdentity();
-    size_t sw=(x2-x1)>5?(x2-x1):5;
-    size_t sh=(y2-y1)>5?(y2-y1):5;
+    unsigned int sw=(x2-x1)>5?(x2-x1):5;
+    unsigned int sh=(y2-y1)>5?(y2-y1):5;
     gluPickMatrix((GLdouble) (x1+x2)/2,(GLdouble) (h - (y1+y2)/2), (GLdouble)sw,(GLdouble)sh, viewport);
     glOrtho(-m_width*(m_eye.y)*0.005f,m_width*(m_eye.y)*0.005f,-m_height*(m_eye.y)*0.005f,m_height*(m_eye.y)*0.005f,m_nearPlane,m_farPlane+100);
     glMatrixMode(GL_MODELVIEW);
@@ -82,7 +82,7 @@ void BackCamera::setCameraForSelectionD(size_t x1,size_t y1,size_t x2,size_t y2,
     gluLookAt(m_eye.x,-990,m_eye.z,m_target.x,m_target.y,m_target.z,m_up.x,m_up.y,m_up.z);
 }
 
-    void BackCamera::setReferenceImage(const char *path,GLuint imageID,Vector &position,size_t width,size_t height)
+    void BackCamera::setReferenceImage(const char *path,GLuint imageID,Vector &position,unsigned int width,unsigned int height)
 {
         if(backReference)
         {
@@ -94,7 +94,7 @@ void BackCamera::setCameraForSelectionD(size_t x1,size_t y1,size_t x2,size_t y2,
         backReference->isShow=true;
 };
 
-void BackCamera::getViewportImage(GLuint &texID,std::string &path,Vector &position,size_t &width,size_t &height)
+void BackCamera::getViewportImage(GLuint &texID,std::string &path,Vector &position,unsigned int &width,unsigned int &height)
 {
     if(backReference)
     {
