+++ Camera/Camera.h
@@ -37,9 +37,9 @@ public:
 
     virtual void disableReference();
 
-    virtual void getViewportImage(GLuint &texID,std::string &path,Vector &position,size_t &width,size_t &height);
+    virtual void getViewportImage(GLuint &texID,std::string &path,Vector &position,unsigned int &width,unsigned int &height);
 
-    virtual void setReferenceImage(const char *path,GLuint imageID,Vector &position,size_t width,size_t height);
+    virtual void setReferenceImage(const char *path,GLuint imageID,Vector &position,unsigned int width,unsigned int height);
 
     virtual void onRotatePress(int x,int y);
 
@@ -53,7 +53,7 @@ public:
 
     virtual void setCamera();
 
-    virtual void setCameraForSelectionD(size_t x1,size_t y1,size_t x2,size_t y2,size_t h);
+    virtual void setCameraForSelectionD(unsigned int x1,unsigned int y1,unsigned int x2,unsigned int y2,unsigned int h);
 
     virtual void setCameraForSelectionS();
