+++ Camera/PerspectiveCamera.cpp
@@ -203,15 +203,15 @@ void PerspectiveCamera::setCamera()
     glLightfv(GL_LIGHT0, GL_POSITION, m_lightPosition);
 }
 
-void PerspectiveCamera::setCameraForSelectionD(size_t x1,size_t y1,size_t x2,size_t y2,size_t h)
+void PerspectiveCamera::setCameraForSelectionD(unsigned int x1,unsigned int y1,unsigned int x2,unsigned int y2,unsigned int h)
 {
     glViewport((GLint)m_startX,(GLint)m_startY,(GLint)m_width,(GLint)m_height);
     GLint viewport[4];
     glGetIntegerv (GL_VIEWPORT, viewport);
     glMatrixMode(GL_PROJECTION);
     glLoadIdentity();
-    size_t sw=(x2-x1)>5?(x2-x1):5;
-    size_t sh=(y2-y1)>5?(y2-y1):5;
+    unsigned int sw=(x2-x1)>5?(x2-x1):5;
+    unsigned int sh=(y2-y1)>5?(y2-y1):5;
     gluPickMatrix((GLdouble) (x1+x2)/2,(GLdouble) (h - (y1+y2)/2), (GLdouble)sw,(GLdouble)sh, viewport);
     gluPerspective(m_angle,m_width/m_height,m_nearPlane,m_farPlane);
     glMatrixMode(GL_MODELVIEW);
