+++ Camera/TopCamera.cpp
@@ -66,15 +66,15 @@ void TopCamera::setCameraForSelectionS()
     glViewport((GLint) m_startX,(GLint) m_startY,(GLint) m_width,(GLint) m_height);
 };
 
-void TopCamera::setCameraForSelectionD(size_t x1,size_t y1,size_t x2,size_t y2,size_t h)
+void TopCamera::setCameraForSelectionD(unsigned int x1,unsigned int y1,unsigned int x2,unsigned int y2,unsigned int h)
 {
     glViewport((GLint)m_startX,(GLint)m_startY,(GLint)m_width,(GLint)m_height);
     GLint viewport[4];
     glGetIntegerv (GL_VIEWPORT, viewport);
     glMatrixMode(GL_PROJECTION);
     glLoadIdentity();
-    size_t sw=(x2-x1)>5?(x2-x1):5;
-    size_t sh=(y2-y1)>5?(y2-y1):5;
+    unsigned int sw=(x2-x1)>5?(x2-x1):5;
+    unsigned int sh=(y2-y1)>5?(y2-y1):5;
     gluPickMatrix((GLdouble) (x1+x2)/2,(GLdouble) (h - (y1+y2)/2), (GLdouble)sw,(GLdouble)sh, viewport);
     glOrtho(-m_width*m_eye.z*0.005f,m_width*m_eye.z*0.005f,-m_height*m_eye.z*0.005f,m_height*m_eye.z*0.005f,m_nearPlane,m_farPlane);
     glMatrixMode(GL_MODELVIEW);
@@ -82,7 +82,7 @@ void TopCamera::setCameraForSelectionD(size_t x1,size_t y1,size_t x2,size_t y2,s
     gluLookAt(m_eye.x,m_eye.y,990,m_target.x,m_target.y,m_target.z,m_up.x,m_up.y,m_up.z);
 };
 
-void TopCamera::setReferenceImage(const char *path,GLuint imageID,Vector &position,size_t width,size_t height)
+void TopCamera::setReferenceImage(const char *path,GLuint imageID,Vector &position,unsigned int width,unsigned int height)
 {
         if(topReference)
         {
@@ -94,7 +94,7 @@ void TopCamera::setReferenceImage(const char *path,GLuint imageID,Vector &positi
         topReference->isShow=true;
 };
 
-void TopCamera::getViewportImage(GLuint &texID,std::string &path,Vector &position,size_t &width,size_t &height)
+void TopCamera::getViewportImage(GLuint &texID,std::string &path,Vector &position,unsigned int &width,unsigned int &height)
 {
     if(topReference)
     {
