+++ History/HistoryLog.cpp
@@ -1,6 +1,6 @@
 #include "HistoryLog.h"
 
-HistoryLog::HistoryLog(LogType::__Enum theType,size_t theTarget):type(theType),target(theTarget)
+HistoryLog::HistoryLog(LogType::__Enum theType,unsigned int theTarget):type(theType),target(theTarget)
 {
 	//printf("f");
 }
